Parallel is not possible due to how postgres manages failures. it invalidates the transaction until commit or rollback which forces one to use auto commit = true, defeating the purpose of parallelism
